Move folders and fix errors from schema change
