+++ src/components/Product.tsx
@@ -1,4 +1,3 @@
-import { Product as ProductType } from '@prisma/client';
 import { Overpass_Mono } from 'next/font/google';
 
 const overpassMono = Overpass_Mono({ subsets: ['latin'] });
@@ -10,9 +9,7 @@ export default function Product({ product }: any) {
         <a href={product.ProductScraper[0].url}>
           <div className='border-l-2 border-l-green-500 bg-green-50 p-3'>
             <img
-              src={
-                product.ProductScraper[0].ProductScraperHistory[0].imageBase64
-              }
+              src={product.ProductScraper[0].ProductScraperHistory[0].imageUrl}
             />
           </div>
         </a>
